chore: Refactoring RowExpanding into folder with separate base, types, and utils
