+++ packages/table-core/src/core/table.ts
@@ -32,7 +32,7 @@ import { ColumnSizing } from '../features/column-sizing/ColumnSizing'
 import { ColumnVisibility } from '../features/column-visibility/ColumnVisibility'
 import { GlobalFaceting } from '../features/global-faceting/GlobalFaceting'
 import { GlobalFiltering } from '../features/global-filtering/GlobalFiltering'
-import { RowExpanding } from '../features/RowExpanding'
+import { RowExpanding } from '../features/row-expanding/RowExpanding'
 import { RowPagination } from '../features/RowPagination'
 import { RowPinning } from '../features/row-pinning/RowPinning'
 import { RowSelection } from '../features/RowSelection'
